refactor: removed duplicate storybook code
